feat/fix: Created separate file to call groq and replicate API for review and related changes
