+++ code-mentor-ai/src/app/api/code-review/route.ts
@@ -1,6 +1,8 @@
 // app/api/code-review/route.ts
 import { NextRequest, NextResponse } from 'next/server';
 import getGroqAiReview from '../groq';
+import getReplicateAiReview from '../replicate';
+
 
 export interface CodeReviewRequest {
     code: string;
@@ -16,8 +18,9 @@ export interface CodeReviewResponse {
     refactoredCode: string;
 }
 
+const CURRENT_AI_MODEL = process.env.CURRENT_AI_MODEL;
 const REPLICATE_API_TOKEN = process.env.REPLICATE_API_TOKEN;
-const MODEL_VERSION_ID = "meta/meta-llama-3-8b-instruct"; // Chosen for cost-effective high-volume use
+
 
 const matchSection = (data: string, labels: string[], regexStr: string, fallback = ''): string => {
     for (const label of labels) {
@@ -27,6 +30,7 @@ const matchSection = (data: string, labels: string[], regexStr: string, fallback
     }
     return fallback;
 };
+
 
 // Parses bullet-pointed list (- item) and numbered list (1. item)
 const parseList = (text: string, isNumbered: boolean = false): string[] => {
@@ -46,6 +50,53 @@ const parseList = (text: string, isNumbered: boolean = false): string[] => {
 
 };
 
+
+const parseGroqAiResponse = async (
+    code: string, language: string, description: string
+): Promise<CodeReviewResponse> => {
+
+    const groqAiReview = await getGroqAiReview(code, language, description);
+
+    // Added **** to mark the end of the review 
+    // (so that refactored code can be parsed easily, as it may have some comments starting with /** */)
+    const resultText = groqAiReview + "\n****";
+
+    const regexStr = '([\\s\\S]*?)(?=\\n*\\s*(\\d+\\.\\s|\\*\\*[\\s]*|$))';
+    return {
+        review: matchSection(resultText, ['Code Review:'], regexStr, "No review provided"),
+        suggestions: parseList(matchSection(resultText, ['Suggestions:'], regexStr), false),
+        potentialBugs: parseList(matchSection(resultText, ['Potential Bugs:'], regexStr), false),
+        securityIssues: parseList(matchSection(resultText, ['Security Issues:'], regexStr), false),
+        refactoredCode: matchSection(
+            resultText, ['Refactored Code:'],
+            '([\\s\\S]*?)(?=\\n*\\s*(\\*\\*\\*\\*))', code
+        ),
+    }
+}
+
+
+const parseReplicateAiResponse = async (
+    code: string, language: string, description: string
+): Promise<CodeReviewResponse | undefined> => {
+
+    const replicateAiResponse = await getReplicateAiReview(code, language, description);
+
+    if (!replicateAiResponse) {
+        return undefined;
+    }
+
+    const resultText = replicateAiResponse + "\n**";
+    const regexStr = '([\\s\\S]*?)(?=\\n*\\s*(\\d+\\.\\s|\\*\\*[\\s]*|$))';
+    return {
+        review: matchSection(resultText, ['Code Review'], regexStr, "No review provided"),
+        suggestions: parseList(matchSection(resultText, ['Suggestions'], regexStr)),
+        potentialBugs: parseList(matchSection(resultText, ['Potential Bugs'], regexStr)),
+        securityIssues: parseList(matchSection(resultText, ['Security Issues'], regexStr)),
+        refactoredCode: matchSection(resultText, ['Refactored Code'], regexStr, code),
+    }
+}
+
+
 export async function getAIReview(
     code: string, language: string, description: string
 ): Promise<CodeReviewResponse> {
@@ -54,95 +105,22 @@ export async function getAIReview(
     }
 
     try {
-        const prompt = `You are an expert code reviewer with deep knowledge of ${language}. Review the following code based on the provided description and return a detailed analysis in the exact format specified below. Be thorough, specific, and actionable in your feedback.
 
-        Language: ${language}
-        Description: ${description}
-        Code:
-        ${code}
+        let res: CodeReviewResponse = {
+            review: "Unable to generate review due to API error",
+            suggestions: [],
+            potentialBugs: [],
+            securityIssues: [],
+            refactoredCode: code,
+        };
 
-        Respond in this exact format:
-        1. Code Review: [Provide a detailed paragraph reviewing the code's quality, structure, and adherence to best practices for ${language}. Mention specific strengths and weaknesses.]
-        2. Suggestions:
-        - [Specific, actionable suggestion 1 with explanation]
-        - [Specific, actionable suggestion 2 with explanation]
-        - [Add more as needed, or leave empty if none]
-        3. Potential Bugs:
-        - [Specific potential bug 1 with line number if applicable and explanation]
-        - [Specific potential bug 2 with line number if applicable and explanation]
-        - [Add more as needed, or leave empty if none]
-        4. Security Issues:
-        - [Specific security issue 1 with explanation and potential impact]
-        - [Specific security issue 2 with explanation and potential impact]
-        - [Add more as needed, or leave empty if none]
-        5. Refactored Code:
-        [Provide the complete refactored code with improvements applied. If no changes are needed, return the original code unchanged. Include comments explaining changes.]`;
-
-        const predictionResponse = await fetch("https://api.replicate.com/v1/predictions", {
-            method: "POST",
-            headers: {
-                "Authorization": `Token ${REPLICATE_API_TOKEN}`,
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify({
-                version: MODEL_VERSION_ID,
-                input: {
-                    prompt,
-                    temperature: 0.2,
-                    top_p: 0.9,
-                    max_new_tokens: 1024
-                }
-            })
-        });
-        let resultText = "";
-        let res: CodeReviewResponse;
-
-        const predictionData = await predictionResponse.json();
-
-        if (predictionData.status == 402) {
+        if (CURRENT_AI_MODEL == "GROQ") {
             // If replicate limit exceeds, switch to groq API
-            const groqAiReview = await getGroqAiReview(code, language, description);
-            resultText += groqAiReview + "\n****";
-            const regexStr = '([\\s\\S]*?)(?=\\n*\\s*(\\d+\\.\\s|\\*\\*[\\s]*|$))';
-            res = {
-                review: matchSection(resultText, ['Code Review:'], regexStr, "No review provided"),
-                suggestions: parseList(matchSection(resultText, ['Suggestions:'], regexStr), false),
-                potentialBugs: parseList(matchSection(resultText, ['Potential Bugs:'], regexStr), false),
-                securityIssues: parseList(matchSection(resultText, ['Security Issues:'], regexStr), false),
-                refactoredCode: matchSection(
-                    resultText, ['Refactored Code:'],
-                    '([\\s\\S]*?)(?=\\n*\\s*(\\*\\*\\*\\*))', code
-                ),
-            }
+            res = await parseGroqAiResponse(code, language, description);
+
         } else {
-            const predictionUrl = predictionData?.urls?.get;
-            while (true) {
-                const res = await fetch(predictionUrl, {
-                    headers: { "Authorization": `Token ${REPLICATE_API_TOKEN}` }
-                });
-                const status = await res.json();
-                if (status.status === "succeeded") {
-                    resultText = "";
-                    for (const output of status.output) {
-                        resultText += output;
-                    }
-                    resultText += '\n**';
-                    break;
-                } else if (status.status === "failed") {
-                    throw new Error("Replicate prediction failed");
-                }
-                await new Promise(r => setTimeout(r, 1000));
-            }
-
-            const regexStr = '([\\s\\S]*?)(?=\\n*\\s*(\\d+\\.\\s|\\*\\*[\\s]*|$))';
-            res = {
-                review: matchSection(resultText, ['Code Review'], regexStr, "No review provided"),
-                suggestions: parseList(matchSection(resultText, ['Suggestions'], regexStr)),
-                potentialBugs: parseList(matchSection(resultText, ['Potential Bugs'], regexStr)),
-                securityIssues: parseList(matchSection(resultText, ['Security Issues'], regexStr)),
-                refactoredCode: matchSection(resultText, ['Refactored Code'], regexStr, code),
-            }
-
+            const response = await parseReplicateAiResponse(code, language, description);
+            res = response ? response : res;
         }
 
         console.log(res);
@@ -159,6 +137,7 @@ export async function getAIReview(
         };
     }
 }
+
 
 export async function POST(request: NextRequest): Promise<NextResponse> {
     try {
